+++ pyi.py
@@ -510,7 +510,7 @@ def _has_bad_hardcoded_returns(
     ):
         return False
 
-    if not _non_kw_only_args_of(method.args):  # weird, but theoretically possible
+    if not method.args.posonlyargs and not method.args.args:  # weird, but theoretically possible
         return False
 
     method_name, returns = method.name, method.returns
@@ -583,13 +583,6 @@ def _is_bad_TypedDict(node: ast.Call) -> bool:
         fieldname.isidentifier() and not iskeyword(fieldname)
         for fieldname in fieldnames
     )
-
-
-def _non_kw_only_args_of(args: ast.arguments) -> list[ast.arg]:
-    """Return a list containing the pos-only args and pos-or-kwd args of `args`"""
-    # pos-only args don't exist on 3.7
-    pos_only_args: list[ast.arg] = getattr(args, "posonlyargs", [])
-    return pos_only_args + args.args
 
 
 def _is_assignment_which_must_have_a_value(
@@ -1581,7 +1574,7 @@ class PyiVisitor(ast.NodeVisitor):
         self, node: ast.FunctionDef | ast.AsyncFunctionDef, method_name: str
     ) -> None:
         all_args = node.args
-        non_kw_only_args = _non_kw_only_args_of(all_args)
+        non_kw_only_args = all_args.posonlyargs + all_args.args
         num_args = len(non_kw_only_args)
         varargs = all_args.vararg
 
@@ -1747,7 +1740,7 @@ class PyiVisitor(ast.NodeVisitor):
         if all_args.kwonlyargs:
             return
 
-        non_kw_only_args = _non_kw_only_args_of(all_args)
+        non_kw_only_args = all_args.posonlyargs + all_args.args
 
         # Raise an error for defining __str__ or __repr__ on a class, but only if:
         # 1). The method is not decorated with @abstractmethod
@@ -1785,15 +1778,12 @@ class PyiVisitor(ast.NodeVisitor):
     ) -> None:
         cleaned_method = deepcopy(node)
         cleaned_method.decorator_list.clear()
-        first_arg = _non_kw_only_args_of(cleaned_method.args)[0]
-        first_arg.annotation = None
+        non_kw_only_args = cleaned_method.args.posonlyargs + cleaned_method.args.args
+        non_kw_only_args[0].annotation = None
         new_syntax = _unparse_func_node(cleaned_method)
         new_syntax = re.sub(rf"\b{typevar_name}\b", "Self", new_syntax)
         self.error(
-            # pass the node for the first argument to `self.error`,
-            # rather than the function node,
-            # as linenos differ in Python 3.7 and 3.8+ for decorated functions
-            node.args.args[0],
+            node,
             Y019.format(typevar_name=typevar_name, new_syntax=new_syntax),
         )
 
@@ -1904,7 +1894,7 @@ class PyiVisitor(ast.NodeVisitor):
             self.generic_visit(node)
 
     def visit_arguments(self, node: ast.arguments) -> None:
-        args = _non_kw_only_args_of(node)
+        args = node.posonlyargs + node.args
         defaults = [None] * (len(args) - len(node.defaults)) + node.defaults
         assert len(args) == len(defaults)
         for arg, default in zip(args, defaults):